+++ concatenator/stitchee.py
@@ -4,32 +4,31 @@ from __future__ import annotations
 
 import logging
 import os
+import shutil
 import time
 from contextlib import ExitStack
 from logging import Logger
+from pathlib import Path
 from warnings import warn
 
 import netCDF4 as nc
 import xarray as xr
 
-from concatenator import GROUP_DELIM
-from concatenator.dimension_cleanup import remove_duplicate_dims
-from concatenator.file_ops import add_label_to_path
-from concatenator.group_handling import (
+import concatenator
+from concatenator.dataset_and_group_handling import (
     flatten_grouped_dataset,
     regroup_flattened_dataset,
+    validate_workable_files,
+)
+from concatenator.dimension_cleanup import remove_duplicate_dims
+from concatenator.file_ops import (
+    add_label_to_path,
+    make_temp_dir_with_input_file_copies,
+    validate_input_path,
+    validate_output_path,
 )
 
 default_logger = logging.getLogger(__name__)
-
-# class netcdfExitStack(ExitStack):
-#     """A context manager that handles netCDF.Dataset.close exceptions."""
-#     def __exit__(self, *args, logger=default_logger, **kwargs):
-#         try:
-#             super().__exit__(*args, **kwargs)
-#         except RuntimeError as err:
-#             if str(err) == "NetCDF: Not a valid ID":
-#                 logger.warning("Tried closing an already closed netCDF.")
 
 
 def stitchee(
@@ -37,35 +36,56 @@ def stitchee(
     output_file: str,
     write_tmp_flat_concatenated: bool = False,
     keep_tmp_files: bool = True,
-    concat_method: str = "xarray-concat",
+    concat_method: str | None = "xarray-concat",
     concat_dim: str = "",
     concat_kwargs: dict | None = None,
     history_to_append: str | None = None,
+    copy_input_files: bool = False,
+    overwrite_output_file: bool = False,
+    group_delimiter: str = "__",
     logger: Logger = default_logger,
 ) -> str:
     """Concatenate netCDF data files along an existing dimension.
 
     Parameters
     ----------
-    files_to_concat : list[str]
-    output_file : str
+    files_to_concat
+        netCDF files to concatenate
+    output_file
+        file path for output file
     write_tmp_flat_concatenated
-    keep_tmp_files : bool
+        whether to save intermediate flattened files or not (default: False).
+    keep_tmp_files
+        whether to keep all temporary files created (default: True).
     concat_method
-    concat_dim : str, optional
+        either "xarray-concat" (default) or "xarray-combine".
+    concat_dim
+        dimension along which to concatenate (default: ""). Not needed is concat_method is "xarray-combine".
     concat_kwargs
+        keyword arguments to pass to xarray.concat or xarray.combine_by_coords (default: None).
     history_to_append
-    logger : logging.Logger
+        json string to append to the history attribute of the concatenated file (default: None).
+    copy_input_files
+        whether to copy input files or not (default: False).
+    overwrite_output_file
+        whether to overwrite output file (default: False).
+    group_delimiter
+        character used to separate groups (default: "__").
+    logger
 
     Returns
     -------
     str
+        path of concatenated file
     """
+    validate_input_path(files_to_concat)
+    concatenator.group_delim = group_delimiter
+
     intermediate_flat_filepaths: list[str] = []
     benchmark_log = {"flattening": 0.0, "concatenating": 0.0, "reconstructing_groups": 0.0}
 
     # Proceed to concatenate only files that are workable (can be opened and are not empty).
-    input_files, num_input_files = _validate_workable_files(files_to_concat, logger)
+    input_files, num_input_files = validate_workable_files(files_to_concat, logger)
 
     # Exit cleanly if no workable netCDF files found.
     if num_input_files < 1:
@@ -76,6 +96,15 @@ def stitchee(
         warn(
             "'concat_dim' was specified, but will not be used because xarray-combine method was "
             "selected."
+        )
+
+    output_file = validate_output_path(output_file, overwrite=overwrite_output_file)
+
+    # If requested, make a temporary directory with new copies of the original input files
+    temporary_dir_to_remove = None
+    if copy_input_files:
+        input_files, temporary_dir_to_remove = make_temp_dir_with_input_file_copies(
+            input_files, Path(output_file)
         )
 
     try:
@@ -107,7 +136,7 @@ def stitchee(
                     decode_coords=False,
                     drop_variables=coord_vars,
                 )
-                first_value = xrds[GROUP_DELIM + concat_dim].values.flatten()[0]
+                first_value = xrds[concatenator.group_delim + concat_dim].values.flatten()[0]
                 concat_dim_order.append(first_value)
 
                 benchmark_log["flattening"] = time.time() - start_time
@@ -141,7 +170,7 @@ def stitchee(
             if concat_method == "xarray-concat":
                 combined_ds = xr.concat(
                     xrdataset_list,
-                    dim=GROUP_DELIM + concat_dim,
+                    dim=concatenator.group_delim + concat_dim,
                     data_vars="minimal",
                     coords="minimal",
                     **concat_kwargs,
@@ -154,7 +183,7 @@ def stitchee(
                     **concat_kwargs,
                 )
             else:
-                raise ValueError("Unexpected concatenation method, <%s>." % concat_method)
+                raise ValueError(f"Unexpected concatenation method, <{concat_method}>.")
 
             benchmark_log["concatenating"] = time.time() - start_time
 
@@ -189,6 +218,8 @@ def stitchee(
                     os.remove(file)
                 if tmp_flat_concatenated_path:
                     os.remove(tmp_flat_concatenated_path)
+                if not keep_tmp_files and temporary_dir_to_remove:
+                    shutil.rmtree(temporary_dir_to_remove)
 
     except Exception as err:
         logger.info("Stitchee encountered an error!")
@@ -196,32 +227,3 @@ def stitchee(
         raise err
 
     return output_file
-
-
-def _validate_workable_files(files_to_concat, logger) -> tuple[list[str], int]:
-    """Remove files from list that are not open-able as netCDF or that are empty."""
-    workable_files = []
-    for file in files_to_concat:
-        try:
-            with nc.Dataset(file, "r") as dataset:
-                is_empty = _is_file_empty(dataset)
-                if is_empty is False:
-                    workable_files.append(file)
-        except OSError:
-            logger.debug("Error opening <%s> as a netCDF dataset. Skipping.", file)
-
-    number_of_workable_files = len(workable_files)
-
-    return workable_files, number_of_workable_files
-
-
-def _is_file_empty(parent_group: nc.Dataset | nc.Group) -> bool:
-    """
-    Function to test if a all variable size in a dataset is 0
-    """
-    for var in parent_group.variables.values():
-        if var.size != 0:
-            return False
-    for child_group in parent_group.groups.values():
-        return _is_file_empty(child_group)
-    return True